Update to sphinx documentation (#3078)
This updates our website to use sphinx for its documentation site. This
includes some new content around the SQL API, as well as our API
Javadoc, which we will be publishing ourselves for the first time in a
while. Publishing happens at the end of the release process (which I've
been able to test by first applying these changes to a test repo of
mine).

---------

Co-authored-by: FDB Relational Layer Authors <foundationdb_ci@apple.com>