Cache FDB packages during the build and avoid checking them in (#3164)
This is to prevent another build from failing to produce a PR on
unsuccessful merge (see #3163). It removes the debian packages from the
working directory, and it also excludes `*.deb` files in the
`.gitignore`, just to be sure. While I was there, I also added artifact
caching so that the directory would be included in the build cache.
There may not be much of a difference, given that the artifacts being
downloaded are hosted on GitHub anyway, but it seemed like a good
practice.