Add a check that makes sure the PRs have the right label
This reads the release notes config to generate a list of expected labels. It then compares the list of labels set on the PR and checks to make sure the PR has one of the labels provided.

Not in this PR: in theory, there are some GraphQL queries we could run so that if we have an associated issue, we could try and guess the label from the issue. This keeps it simple, though, and just makes sure that the label has been set. If the check fails, it should be a pretty straightforward and fast thing to fix.