Enable a few tests where the force continuation option does not fail anymore (#3176)
Testing with forced continuations between `main` and 4.0.559.6 showed
that some tests that were excluded because of `FORCED_CONTINUATIONS` can
now be enabled.
The tests that were enabled pass on the main -> `4.1.6.0` mixed-mode
tests, though they do fail when run against `4.0.559.6`. This is how we
want it, since it would flag them as needing attention.