Generate release notes automatically from pull requests (#3174)
This alleviates two issues:
1. Developers routinely hit conflicts on their pull requests, requiring
merging. The original code had attempted to alleviate this by having
multiple different items that could be replaced by category, but it was
insufficient.
2. In part due to the above, presumably, a lot of release notes in the
past were sparse, by generating it automatically from the PR, this will
be much harder.

I considered just using GitHub's `generate-notes` functionality, but
that is rather limited. This change doesn't really expand on using that
yet, but puts us in a better place to do so.

This also updates the release workflow to use the new script, and
updates the release notes so that they have the new format for every
version since `4.0.559.0`. This also means the release notes now include
the patch releases of `4.0.559.*`. I also did some manual additions to
the release notes.

---------

Co-authored-by: Alec Grieser <alloc@apple.com>