Add version information to yaml test exceptions (#3189)
This does the following:
1. It introduces a `YamlConnection` interface that wraps the underlying
connection.
2. This makes it easier to have the connection expose what versions it
is testing against

The issue we found is that when running in mixed-mode it runs the same
query multiple times in parallel (as intended, I just didn't realize).
This makes it incredibly hard to tell if a failing test is failing
because it fails when going from the old version to the new version, or
vice-versa.