Fix instability in the query planner (#3185)
This fixes a number of instabilities found in Cascades when planning
certain type of queries that end up generating a relatively large number
of variations that might end up storing actual duplicates in the
memoization structure (which is permissible).

Essentially, the following issues were detected and resolved in this PR:
1. **Fix instability in hash calculation**: Certain expressions had
their hash unstable, because it includes hashing an `enum` member, which
is [known to be
unstable](https://bugs.openjdk.org/browse/JDK-8050217?page=com.atlassian.jira.plugin.system.issuetabpanels%3Aall-tabpanel).
This issue made it very difficult to compare planner traces from
different runs, since a planner trace might include the hash codes of
some expressions, and the hash code could change across different JVMs,
causing the comparator to report false negatives.
2. **Set instantiation with deterministic traversal order**: Certain
instantiations of `Maps.newIdentityHashMap` and `Sets.newHashSet` are
replaced with `LinkedIdentityMap`, resp. `LinkedIdentitySet`.
3. **Fix instability detected in Guava `StandardMutableNetwork`**: We
use Guava [`Network`
](https://guava.dev/releases/20.0/api/docs/com/google/common/graph/Network.html)
for the planner memoization structure, internally, the network uses a
hash map to store the parallel edges between network vertices, this is
exactly why we were getting the alternatives in different order as
explained above. Due to the way the `Network` code is written, it is not
possible to pass the (type of the) map to the builder, therefore, I went
ahead and copied enough of the code to be able to modify the map
construction and make it stable, which effectively fixed the issue, but
ideally, Guava should make it possible to control how these maps are
created, making it easier to use and reason about the code.
This fixes #3191.

---------

Co-authored-by: Normen Seemann <nseemann@apple.com>