Add protection against test running into an infinite loop (#3186)
Trying to improve the test experience, abort tests in case they are
running into infinite loops when processing continuations.
This assumes that the test data will not produce more than a maximum
number of rows (currently set to 100), and aborts the continuation loop
after that number.
Note that the non-forced-continuation loop does not need that protection
since it will require a "result" line in the yaml to trigger a
continuation , and those are bound by a finite number.