Run nightly variations of yaml-tests in nightly workflow (#3202)
The configuration that existed for yaml-tests used a bunch of
environment variables that we aren't setting. Change it to use the
project properties that we do set.
I removed the control for the number of threads, as I felt it would be
good to simplify the number of controls.
It also always sets and logs the seed used during tests.
Since the yaml tests (currently) are very quick, I set the iterations to
be 50x the iterations used elsewhere, but we can control it manually.

Also, to make it easier to run the nightly workflow, I set it to also
have `workflow_dispatch`