Adds commands to the YAML framework to assert on changing query behavior (#3188)
This adds additional arguments to the YAML test framework to assert on
changing query behavior. For example, if a query used to return fewer
columns and that information is baked into a plan, we might expect the
configuration to look something like:

```yaml
-
  - query: SELECT something()
  - initialVersionLessThan: 4.2.0.0
  - result: [{a: 1, b: 2}, {a: 3, b: 4}]
  - initialVersionAtLeast: 4.2.0.0
  - result: [{a: 1, b: 2, c: "foo"}, {a: 3, b: 4, c: "bar"}]
```

The way to read this is that after the results following
`initialVersionLessThan: 4.2.0.0` should be returned as long as the
query starts on a version less than `4.2.0.0`. If we having
continuations (like if we are running in `force_continuation` mode),
then we get coverage of what happens when we send the continuation up to
a new version, which may be 4.2.0.0+. Likewise, the latter result
(returning more columns) is expected even if the continuation is sent
back to a pre-4.2.0.0 version.

In this PR, I've also updated the YAML tests that were resulting in
mixed-mode failures between 4.1.6.0 and earlier versions so that it now
takes the version into account. This shows the explicit behavior change
that we are now asserting on that previously would have resulted in
queries that were only run on the new version.

This resolves #3114.

---------

Co-authored-by: Scott Dugas <scott.dugas@gmail.com>