Serialize missing field `createDefaultIfEmpty` into streaming aggregate plan continuations (#3211)
There is missing state that was introduced to the `AggregateCursor` and
its accompanying plan in #3092. However, this state was not added to the
continuation, which results in unexpected behavior when we try to
deserialize the plan's continuation (effectively, it reverts to the old
behavior of the plan operator). We were already correctly
_de_serializing this information, so the upshot is that any plan that is
started after this current version should exhibit the correct behavior,
but we get incorrect behavior when a plan is started on an older
version.

This fixes #3096. I was able to get test coverage of this part of the
code by enabling `FORCE_CONTINUATIONS` mode on `aggregateIndexTests`.