Fix continuation deserialization errors by deserializing streaming aggregate plans in the correct order (#3215)
The error we were seeing was that for certain plans, we were getting
errors trying to deserialize the elements of a
`FDBStreamingAggregatePlan` continuation. In particular, we were seeing
the type's reference before the type's definition. This was because the
serialization and deserialization of the plan happened in different
orders, which meant that at deserialization time, we were no longer
guaranteed to come across the definition first. I've kept the order of
serialization the same, and then updated the deserialization code path
to align it with existing continuations that are out in the wild.

This fixes #3214. I've enabled FORCE_CONTINUATIONS mode on that test,
though it mostly just works on the embedded vs. current configuration
(which is enough to surface the bug). More test coverage is added by
having the `StreamingAggregate` plans in one of the Record Layer tests
all go through `verifySerialization`, which previously wasn't being done
on those plans.