Prevent caching `INSERT` statements (#3212)
This disables adding `INSERT` statements to the plan cache. The main
reason is related to potentially generating very large
`QueryPhysicalPlan` objects due to the way we capture individual
`INSERT` values and generate potentially massive `QueryPlanConstraint`
to verify their type (unnecessarily), in addition to other overheads
such as generating massive amount of redundant `Type` objects and `RCV`s
if the insert is touching a very wide table with hundreds of rows.

This resolves #3210