`RecordQueryExplodePlan` plan adheres to `skip` and `limit` requirements (#3230)
The `RecordQueryExplodePlan` plan is, so far, only used internally to
represent a transposition of constant arrays such as e.g. `in` values as
a leg of an in-join plan, and as long as the upper operator honors the
skip and limit requirements it was probably ok to have this operator
ignore these parameters internally. In fact, it is somewhat common for
the upper operator to remove these requirements before proceeding to
execute the inner operator.

I came across this issue when implementing table-functions, one example
of these functions is to directly select from `values` that translates
to a projection from this operator at runtime. In other words, it
suddenly became a top-level operator, and therefore, it must respect
skip and limit to align with the execution behavior of other operators.

This PR explicitly sets the `limit` and `skip` on the resulting
`RecordCursor` of the `RecordQueryExplodePlan`, it also adds a number of
unit tests.

It resolves #3229.