`EXISTS` queries now honor continuation from previous runs to avoid infinite loops (#3221)
The `RecordQueryFirstOrDefaultPlan` constructs a special cursor which
has cardinality at most one. The continuation it returns back to the
user is not based on the inner cursor but only on whether the cursor has
returned any data. Previously, we were taking the `FutureCursor`'s
continuation (which is always `\x00`) and giving it back to the `inner`
plan, which would generally result in the plan resuming from the
beginning. That is not the appropriate way to interpret the
continuation, so this updates the logic to instead skip creating the
inner cursor if we get a non-beginning continuation, which is in line
with what the `FutureCursor`'s continuation means.

This fixes #3219.