Create a yaml test config that runs two servers with the same version (#3217)
This PR adds a test configuration that runs two servers of the same
version (no embedded in the loop).
In order to do this, a slight refactoring of the config structure was
done, moving all the `create*ConnectionFactory` into a base class.
A new Gradle test config was added so that one can run the new config
from the IDE. No automated way to run the config is being introduced
(this can be added in the future as needed).
Running all the tests (this downloads all recent external servers and
runs a single config on each) takes about 4 minutes. This seems OK, but
if needed we can reduce the scope to one server only.