Fix deserialization of continuations for queries that contain LIKE (#3207)
Previously, it would parse the query pattern as the escape character,
making continuations useless.
i.e. `LIKE 'XYZ' ESCAPE 'XYZ'`, regardless of what the original query
had for the escape character, if it had one.

Because I was curious what happens if the query pattern was a single
character, I added some additional tests, and ran them with the old code
to see if it would break in different ways. It didn't, I think in part
because: `LIKE 'Z' ESCAPE 'Z'` is the same as `LIKE 'Z'`. I don't know
if that is correct behavior, but it is what we do.
Even though these tests didn't expose other issues, I felt like they
were worth keeping.

This Resolves: #3099