Enable force continuations mode on all `YamlIntegrationTest`s (#3228)
This updates the `YamlIntegrationTest`s so that they all run in "force
continuations" mode. That means we get additional insight into how
continuation handling works in those cases.

This mostly accomplishes #3204. There are a few lingering queries that
have not been moved over due to: #3206. In addition, several tests only
run with force continuations on _newer_ versions. However, the majority
of queries have at least some coverage of up-level and down-level
testing for continuations to cover the issues identified as a part of
this process.

---------

Co-authored-by: ohad <ozeliger@apple.com>
Co-authored-by: Scott Dugas <sdugas@apple.com>
Co-authored-by: Scott Dugas <scott.dugas@gmail.com>