Make test output in gradle much more compact
I've been looking at gradle outputs in builds a bunch recently
and the existing format means that the logs are very large.
This changes it to one line per successful test, instead of 3.
Since it prints when it starts, and flushes, you can see what
test is taking a long time when that happens.
Since each test is only one line of text, the blank line seems
completely unnecessary now.