Run Mixed Mode Tests during the release parallel to the other tests (#3233)
This changes our release build to run the MixedModeTest in parallel with
the other tests so that it can populate the release notes at that time.
One key advantage of this is the documentation that sphinx publishes
will have this information.

This also changes the release build so that it is usable for both patch
and non-patch builds, and figures that out depending on the target
branch.

Note: there are a lot of commits here, I don't think I recommend going
through this change commit-by-commit, there was a lot of trial-and-error
to get here.