Queries using versions now use a serialization that requires 4.1
In #2943, a new serialization of the `VersionValue` was introduced, though we had to modify it in #3179 so to avoid deserialization problems when serializing and deserializing continuations across 4.0 and 4.1. With 4.2, we want to start using the new serialization path to serialize `VersionValue` children more completely.

This also bumps the minor version to 4.2. This resolves #3203.