Add missing explicit dependencies for gradle tasks (#3250)
The JMH one is a little surprising to me, as that is just a plugin.

I basically just followed what gradle printed in the warnings. Also, I
validated this locally running each task with a clean, e.g. ./gradlew
clean :fdb-relational-cli:jar --warning-mode all on main, it fails
because files don't exst, but on this branch they succeed.

This is part of the work to allow upgrading to gradle 8:
https://github.com/FoundationDB/fdb-record-layer/issues/3242