Allow having YamlTestExtension include method name in test names
When publishing to JUnit xml style files there is no hierarchy, so
some tools will just show the class name and description, making it
hard to tell what tests were run, or failed. If YamlTestExtension
is used in an environment where that is the case this can be used
to have the output include the method name. For example:
- showcasingTests(Embedded)
- showcasingTests(MultiServer (Embedded then !current_version))
...

Instead of
- YamlIntegrationTests
  - showcasingTests(Runner)
    - Embedded
    - MultiServer (Embedded then !current_version)

I thought about making this configurable via system property, but
I think having it configured by a constructor parameter aligns more
with PRs #3252 and #3251