Remove unused build files (#3257)
We have a number of unused build files which are no longer in use. This
gets rid of them to simplify things and remove unused code. They are:

* `build.py`. This was replaced a while ago with direct calls to
`./gradlew` or with actions like the `acctions/run-gradle` action. It
used to be there to invoke the build correctly for publishing or for
running PRB, but none of the current build or prb functions use it.
* Several Dockerfiles and a `docker-compose.yaml`. These used to form
the basis of a Docker network that could run the tests if need be, but
we've moved away from explicitly specifying a build and instead setting
up the environment in our build config. There's possibly something to be
said about being able to run the server and tests in a containerized
environment, but it's a little bit more involved than one would hope due
to the FDB C client.
* `build/fdb_create_cluster_file.bash`. This has been replaced by
`actions/setup-fdb/action.yml` within the GitHub workflows. It does
mainly the same thing, including making sure there's a cluster file
defined.
* `build/fdb_docker_start.bash`. This was used to start the server
within the Docker files, but it's not used within
`actions/setup-fdb/action.yml`.
* 'build/update_release_notes.bash` and `build/update_release_notes.py`.
These used to update `docs/sphinx/source/ReleaseNotes.md`, but that's a
job now accomplished by `build/create_release_notes.py`. This is updated
in a very different way now, and it doesn't make as much sense to keep
it around.

I've gone through and validated that none of the files referenced here
have any remaining references in this repo (and all of the files in the
`build` directory do).