Remove Artifactory publishing logic from gradle build (#3256)
We no longer need the artifactory publishing logic that we used to use
for publishing in the gradle build as new releases are no longer being
published there. This removes the now dead code from the build, which
means we no longer need to include the artficatory plugin in our
build-time dependencies, which is nice.

There are a few other build files that appear to be dead code that we
can remove, but this is more targeted. I've validated that I can still
publish to maven local with this change, and all of the logic seems to
be pretty specific to actually running an artifactory build, so this
shouldn't disrupt our other build procedures, but I guess one never
knows.

This resolves #3249.