Queries using versions now use a serialization that requires 4.1 (#3245)
In #2943, a new serialization of the `VersionValue` was introduced,
though we had to modify it in #3179 so to avoid deserialization problems
when serializing and deserializing continuations across 4.0 and 4.1.
With 4.2, we want to start using the new serialization path to serialize
`VersionValue` children more completely.

The mixed mode testing in PRB validate this change is compatible with
4.1.9.0. I've also run the full mixed mode battery for the
`versionsTest` tests, and the serialization is compatible between this
change and all 4.1 versions, as well as all 4.0 versions after
4.0.562.0, which is in line with what we'd expect.

This resolves #3203.