Remove a bunch of references to @ExcludeFromJacocoGeneratedReport (#3269)
This marks a bunch of methods to be included in our code coverage
report.

I went through all the references, and I believe I left 3 categories of
things that are excluded:
1. Methods that are simply some variant of throw OperationNotSupported.
We have a lot of these, because we override parts of JDBC, but haven't
implemented all of it yet.
2. There's a handful of Hollow or NoOp classes. I want to investigate
these further, and validate whether we should (a) remove them or (b) at
least have tests that have them where there supposed to go
3. `PlannerDebuggerCommandHandler` -- I believe this command starts the
planner repl, so that you can step through the cascades planner
execution. Given that the planner repl is entirely internal, and
interactive, I feel like this is ok.
4. `LaunchSQLLine` - as noted in the comment on the class, this is just
a little helper to make it easier to debug with sqlline. I will add
though, that I removed some of the exceptions for things used by
SQLLine, so we may want tests that test SQLLine integration, at which
point we could debug from those tests, and thus delete `LaunchSQLLine`
entirely.