Lucene: AnalyzerChooser no longer takes the text when choosing an analyzer (#2994)
This removes the option for having `AnalyzerChooser` pick the analyzer
based on the contents of the query.
This moves it so that the `FDBDirectoryManager` creates one
`LuceneAnalyzerWrapper`, and preserves it for the transaction.
Eventually this will allow `FDBDirectoryWrapper` to have the writer be
lazy but final, but I didn't bring that change into here, isolating this
to the breaking change.

This Resolves #2993