Add cleanup to list of potential labels for PRs (#3268)
I created a PR to cleanup some whitespace, and there was no good label
for it. I saw that we already had a
https://github.com/FoundationDB/fdb-record-layer/labels/cleanup label,
so adding that to the Build/Test/Documentation group in the release
notes seems like a good idea to me.

The description for
https://github.com/FoundationDB/fdb-record-layer/labels/cleanup is
`Stylistic code cleanup`, but I'm not sure if we should have a more
descriptive label, to avoid other types of cleanup, such as deleting
unused, private code.