When downloading old servers, only cache latest version for 10 minutes
The 4.2.3.0 release did not do a mixed-mode test against 4.2.2.0
because that was released less than 24 hours ago. This changes it
so that it refreshes the dynamic version after 10 minutes.
This does mean that builds will have to fetch the latest metadata
more often, including locally.

If this becomes a burden, we could try to add something to our
release process to ensure that after a release we insert update
the cached version, but that seems trickier to manage, and we'd
only really be able to test it by doing releases within 24 hours
of each other.