Fix wasNull state across toString() calls in RowStruct (#3276)
This PR fixes the bug related to `wasNull` state not being preserved
across `toString()` calls in a `RowStruct`. For instance,

If a struct is defined as `{a -> null, b -> 1}`, `getObject(1)` should
set the `wasNull` to `true` which should be uneffected if a call to
`toString()` is made.

Also adds tests around the same.