Added workflow for creating branches for patch releases (#3192)
This creates a new GitHub action to create a release branch from which
additional patch releases can be generated. It takes a version (e.g.,
`4.1.8`) and produces a new branch off of the base tag (so
`4.1.8-release` off of tag `4.1.8.0`). This allows a maintainer to
create a patch branch given our relatively restrictive branch protection
rules, which generally do not allow for branch creation.