Support SQL Table Valued Functions (#3280)
This provides support to table-valued-functions, it introduces the
necessary framework support across different components: parsing,
semantic analysis, plan generation, planning, and execution, to
pre-defined table-valued functions.

Two pre-defined table-functions are introduced in this PR:

#### 1. `values` function:
Most major SQL vendors support this function, the relational engine now
supports this function as well:
```sql
select * from values (1, 'a', 3.0), (4, 'b', 4.5);
-- returns two rows:
-- 1, 'a', 3.0
-- 4, 'b', 4.5
```
It also enables naming the resulting table along with its columns:
```sql
select * from values (1, 2.0, (3, 4, 'foo')), (10, 90.2, (5, 6.0, 'bar')) as A(B, C, W(X, Y, Z))
-- returns two rows with the following metadata:
-- B  | C    | W(X, Y, Z)
-- 1  | 2.0  | (3, 4.0, 'foo')
-- 10 | 90.2 | (5, 6.0, 'bar')
```
Notice how the value `4` in the first row's nested structure `W.Y` was
promoted to `4.0`, this is because this function analyzes all the rows,
and for each leaf in every (nested) tuple, it looks for a type that
accommodates for all the individual values of the leaf across all rows
(max type) and injects type promotions if necessary.
Also note that the type aliasing works with nested array and structs, to
some degrees addresses some of the issues reported in #3231, and in the
future, the new logic can be leveraged to resolve that ticket as well.

#### 2. `range` function:
This is also a very common function, it returns a range between:
- given start (inclusive), defaulting to `0`.
- given end (exclusive).
- optional step size.
- 
```sql
select ID from range (1, 5)
-- returns 4 rows
-- 1
-- 2
-- 3
-- 4
```
The implementation of this TVF has streaming semantics. In other words,
it does not pre-materialize all the values in the range, but rather
return them on-demand.

With this PR, we can hopefully start adding more advanced TVFs easily.

This resolves #3282.