Synthetic record orphan policy (#3271)
Add support for `orphanPolicy` when loading synthetic records. This is
necessary for the ability to scrub indexes with synthetic records as
dangling index entries, by definition, have no records to back them up.
As part of this PR we've also disabled the repair of synthetic records
as part of the scrubbing until we test this feature more thoroughly.
Also of note: Implementors of the `EXPERIMENTAL` `SyntheticRecordType`
interface will see a change that requires the implementation of another
method (or add the new parameter to the existing one).
As far as the "breaking change" label: The scrubbing for dangling index
entries now uses a `RETURN` policy (default was `ERROR`) which means
that previous requests that used to fail now pass (if the record was
missing). Also, any request to scrub synthetic record with repair `on`
will now fail.