Rollback UUID support in SQL Layer (#3293)
This PR reverts the SQL layer support for UUID that was introduced in
#3198. However, the Record Layer support for UUID stays as it is. This
makes this compatible with releases < `4.2.3.0` which was required to
maintain plan continuation compatibility for plans serialized with
non-UUID Type.

Also adds a yaml-test for explicitly defining the RecordMetadata using
proto, and using the `TupleField.UUID` message in it. Owing to
difference in behavior in how this and versions (4.2.3.0 and 4.2.4.0)
interprets `TupleField.UUID`, this PR also blocklists these versions
from yaml-testing.

fixes: #3285