Fix connection name in Multiserver configs in YAML-tests (#3301)
While testing in YAML framework, I realized that a multiserver config
like `MultiServer (4.0.575.0 then Embedded)` is actually connecting to
4.0.575.0 external server and JDBC in-process !current_version (instead
of Embedded). This PR fixes this naming so as to avoid ambiguity.

Also, `MultiServerConnection`s, while alternating their connections,
logs operations like `Sending operation createStatement to connection 0`
meaning, it is sending the current request to first (default, index 0)
connection which is not super clear when debugging issues related to
compatibility. This PR changes that log to explicitly specify the
connection name.