Update a large portion of the code to use FormatVersion
This changes the constants in FDBRecordStore to call through to the
enum. It also adds setters on the builders, and deprecates the old
ones. With this change, I also went through and changed all the
tests to use the enum rather than the constants.

This leaves a couple places untouched:
1. The code within FDBRecordStore still uses the constants
2. getFormatVersion() still returns the integer
3. Code interacting with the result of getFormatVersion() also
references the comments