Streaming aggregate cursor now requires 4.1.9.0 or later for continuation deserialization (#3246)
This cleans up the `StreamingAggregateCursor` and its helper classes to
remove the `createDefaultIfEmpty` option. We'd been carrying that option
around as we introduced that field in #3092 and wanted to preserve
behavior when upgrading from older versions. The intention had been that
all new plans would set the option starting in 4.1, but a bug (fixed in
#3211; see #3096) means that we didn't actually enable it until 4.1.9.0.
That means that after this change, we'd require 4.1.9.0 or newer in
order to safely continue these queries. In theory, we could wait with
this change, but 4.1.9.0 has enough fixes that I actually think our
recommendation should be that anyone upgrading from 4.0 or below go
straight to 4.1.9.0, and then they can proceed safely to a newer version
with this change.

I was able to validate that this change is compatible with 4.1.9.0 via
the cross-version tests run during PRB. When I ran the full mixed mode
tests with the `aggregate-index-tests.yamsql`, I found that only 4.1.9.0
worked, which is expected.

This resolves #3107.