Transition FormatVersion from constants to an enum (#3304)
The FormatVersion was just an integers, with an associated set of
constants, of the form: `FDBRecordStore.*_FORMAT_VERSION`. This PR
starts the migration of the code to a new enum: `FormatVersion`, and
introduces setters and getters that align with this.
This leaves around the existing constants, and methods to avoid being a
breaking change, but they are all now flagged with
`@API(API.Status.DEPRECATED)` and will be removed in the near future.

I did not replace the integer usages in the internals of the record
store, because I felt that code was complicated enough that it would be
worth focusing on independently. I did change everything else to use the
new code.

This also adds a bunch of documentation, and thus resolves: #710.
I also added a bunch of documentation to the `DatastoreInfo`, because a
lot of the details of that are tightly coupled with the FormatVersion.

Note: This is somewhat related to
https://github.com/FoundationDB/fdb-record-layer/pull/3309 as some of
the new paths will also throw an exception on invalid states, whereas
the integer versions would not.

---------

Co-authored-by: Alec Grieser <alloc@apple.com>