Add support for record validation (#3314)
This PR adds two new facilities in support of record validation:
- Scan record keys in a store
- Validate record (no repair yet)

Record validation is meant to detect cases of missing KVs in the
database and (optionally) repair them.
The combination of the two facilities mentioned above would allow for
the scanning of a store that may have suffered data missing and repair
records such that the store is in consistent state.
Constraints of the solution:
- Missing KV only (no mis-ordered splits or corrupt keys)
- Format versions of 3 and higher (no upgrades of data from one format
to another tested)
- With and without split long record option
- With and without inline versions

Resolves #3313